+++ x-pack/elastic-agent/pkg/agent/application/managed_mode.go
@@ -17,6 +17,7 @@ import (
 	"github.com/elastic/beats/v7/x-pack/elastic-agent/pkg/agent/application/info"
 	"github.com/elastic/beats/v7/x-pack/elastic-agent/pkg/agent/application/paths"
 	"github.com/elastic/beats/v7/x-pack/elastic-agent/pkg/agent/application/pipeline"
+	"github.com/elastic/beats/v7/x-pack/elastic-agent/pkg/agent/application/pipeline/emitter"
 	"github.com/elastic/beats/v7/x-pack/elastic-agent/pkg/agent/application/pipeline/router"
 	"github.com/elastic/beats/v7/x-pack/elastic-agent/pkg/agent/application/pipeline/stream"
 	"github.com/elastic/beats/v7/x-pack/elastic-agent/pkg/agent/application/upgrade"
@@ -137,7 +138,7 @@ func newManaged(
 		return nil, errors.New(err, "failed to initialize composable controller")
 	}
 
-	emit, err := emitter(
+	emit, err := emitter.New(
 		managedApplication.bgContext,
 		log,
 		agentInfo,
