+++ x-pack/elastic-agent/pkg/agent/application/pipeline/emitter/emitter_test.go
@@ -2,4 +2,4 @@
 // or more contributor license agreements. Licensed under the Elastic License;
 // you may not use this file except in compliance with the Elastic License.
 
-package application
+package emitter
