+++ x-pack/elastic-agent/pkg/config/operations/inspector.go
@@ -8,6 +8,7 @@ import (
 	"fmt"
 
 	"github.com/elastic/beats/v7/libbeat/logp"
+	"github.com/elastic/beats/v7/x-pack/elastic-agent/pkg/agent/application/info"
 	"github.com/elastic/beats/v7/x-pack/elastic-agent/pkg/agent/application/paths"
 	"github.com/elastic/beats/v7/x-pack/elastic-agent/pkg/agent/configuration"
 	"github.com/elastic/beats/v7/x-pack/elastic-agent/pkg/agent/errors"
@@ -73,7 +74,7 @@ func loadConfig(configPath string) (*config.Config, error) {
 	// merge local configuration and configuration persisted from fleet.
 	rawConfig.Merge(config)
 
-	if err := InjectAgentConfig(rawConfig); err != nil {
+	if err := info.InjectAgentConfig(rawConfig); err != nil {
 		return nil, err
 	}
